move manageiq_entities to module utils
